refactor: Improved code structure and maintainability
